Move code unused on some architectures
Should not change behavior.
