+++ pkg/chrootarchive/jsoniter.go
@@ -1,3 +1,6 @@
+//go:build !windows && !darwin
+// +build !windows,!darwin
+
 package chrootarchive
 
 import jsoniter "github.com/json-iterator/go"
