+++ pkg/unshare/unshare.go
@@ -5,18 +5,12 @@ import (
 	"os"
 	"os/user"
 	"sync"
-
-	"github.com/sirupsen/logrus"
 )
 
 var (
 	homeDirOnce sync.Once
 	homeDirErr  error
 	homeDir     string
-
-	hasCapSysAdminOnce sync.Once
-	hasCapSysAdminRet  bool
-	hasCapSysAdminErr  error
 )
 
 // HomeDir returns the home directory for the current user.
@@ -35,15 +29,4 @@ func HomeDir() (string, error) {
 		homeDir, homeDirErr = home, nil
 	})
 	return homeDir, homeDirErr
-}
-
-func bailOnError(err error, format string, a ...interface{}) { // nolint: golint,goprintffuncname
-	if err != nil {
-		if format != "" {
-			logrus.Errorf("%s: %v", fmt.Sprintf(format, a...), err)
-		} else {
-			logrus.Errorf("%v", err)
-		}
-		os.Exit(1)
-	}
 }
